[GITHUB-9] Separates auth messages for shipping
This change adds shipping config and a dedicated template for
shipping auth and authpriv messages.

During this change it was noticed that the current syslog
shipping config pushed auth messages despite it being intended
to not do so, this update fixes that with a version bump to
indicate the change.

Finally the Vagrant setup was not reading the ansible.cfg file
due to a missing symlink which has been added in.